feat(questions): add GroupButton and GroupSelectionHeader components for improved group selection UI
